Multiplier stage in float value. (value 0.1)
I had problems with the multiplier below 0 and made some changes. If it is incorrect, I ask you to change it.

+++ src/game.h
@@ -457,7 +457,7 @@ class Game
 		}
 
 		bool loadExperienceStages();
-		uint64_t getExperienceStage(uint32_t level);
+		float getExperienceStage(uint32_t level);
 
 		void loadMotdNum();
 		void saveMotdNum() const;
@@ -520,7 +520,7 @@ class Game
 		std::unordered_map<std::string, Player*> mappedPlayerNames;
 		std::unordered_map<uint32_t, Guild*> guilds;
 		std::unordered_map<uint16_t, Item*> uniqueItems;
-		std::map<uint32_t, uint32_t> stages;
+		std::map<float, float> stages;
 
 		std::list<Item*> decayItems[EVENT_DECAY_BUCKETS];
 		std::list<Creature*> checkCreatureLists[EVENT_CREATURECOUNT];
